demo of elevated memory use with rlang 35e87908
On a Bionic (18.04.6) EC2 instance with R 4.1.3, checks on the styler
tarball (1.9.1) are exhausting memory with rlang v1.1.1.  The issue
bisects to rlang 35e879084 (Bump deprecations, 2023-02-24).

This is an attempt to reproduce the issue with the Bionic
rstudio/r-base image.  It looks like that commit does result in a
large increase of the used memory, although it doesn't seem to be
getting to the extreme levels that were need to on the EC2 instance
(21 GB was the largest I tested on).